Rename redux folder
