+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import store from './redux2/configureStore';
+import store from './redux/configureStore';
 
 import Books from './routes/Books';
 import Category from './routes/Category';
